Fix segmentation fault when close the app
A segmentation fault occours when the application is closed. This
happen because the chatbot object is deleted twice in different
class destructors. To fix that, this commit convert some raw
pointers to std::shared_ptr pointers. Thus, multiples objects can
have a reference to the same chatbot object and the object will be
deleted correctly automatically.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>